+++ src/Promise.php
@@ -4,6 +4,7 @@ namespace Http\Adapter\Guzzle6;
 
 use GuzzleHttp\Exception as GuzzleExceptions;
 use GuzzleHttp\Promise\PromiseInterface;
+use Http\Adapter\Guzzle6\Exception\UnexpectedValueException;
 use Http\Client\Exception as HttplugException;
 use Http\Promise\Promise as HttpPromise;
 use Psr\Http\Message\RequestInterface;
@@ -62,9 +63,9 @@ final class Promise implements HttpPromise
             } elseif ($reason instanceof GuzzleExceptions\GuzzleException) {
                 $this->exception = $this->handleException($reason, $request);
             } elseif ($reason instanceof \Throwable) {
-                $this->exception = new \RuntimeException('Invalid exception returned from Guzzle6', 0, $reason);
+                $this->exception = new HttplugException\TransferException('Invalid exception returned from Guzzle6', 0, $reason);
             } else {
-                $this->exception = new \UnexpectedValueException('Reason returned from Guzzle6 must be an Exception');
+                $this->exception = new UnexpectedValueException('Reason returned from Guzzle6 must be an Exception');
             }
 
             throw $this->exception;